requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Pillow==10.4.0
SQLAlchemy==2.0.44
Werkzeug==3.1.3


